# Optmize path volatility tracker