--- a/lib/ui/volatile_path_tracker.h
+++ b/lib/ui/volatile_path_tracker.h
@@ -7,7 +7,7 @@
 
 #include <deque>
 #include <mutex>
-#include <set>
+#include <vector>
 
 #include "flutter/fml/macros.h"
 #include "flutter/fml/task_runner.h"
@@ -33,6 +33,7 @@ class VolatilePathTracker {
   /// The fields of this struct must only accessed on the UI task runner.
   struct TrackedPath {
     bool tracking_volatility = false;
+    bool erased = false;
     int frame_count = 0;
     SkPath path;
   };
@@ -66,13 +67,9 @@ class VolatilePathTracker {
 
  private:
   fml::RefPtr<fml::TaskRunner> ui_task_runner_;
-  std::atomic_bool needs_drain_ = false;
-  std::mutex paths_to_remove_mutex_;
-  std::deque<std::shared_ptr<TrackedPath>> paths_to_remove_;
-  std::set<std::shared_ptr<TrackedPath>> paths_;
+  std::mutex paths_mutex_;
+  std::vector<std::shared_ptr<TrackedPath>> paths_;
   bool enabled_ = true;
-
-  void Drain();
 
   FML_DISALLOW_COPY_AND_ASSIGN(VolatilePathTracker);
 };