--- a/lib/ui/volatile_path_tracker.h
+++ b/lib/ui/volatile_path_tracker.h
@@ -6,6 +6,7 @@
 #define FLUTTER_LIB_VOLATILE_PATH_TRACKER_H_
 
 #include <deque>
+#include <memory>
 #include <mutex>
 #include <vector>
 
@@ -15,6 +16,10 @@
 #include "third_party/skia/include/core/SkPath.h"
 
 namespace flutter {
+
+namespace testing {
+class ShellTest;
+}  // namespace testing
 
 /// A cache for paths drawn from dart:ui.
 ///
@@ -33,7 +38,6 @@ class VolatilePathTracker {
   /// The fields of this struct must only accessed on the UI task runner.
   struct TrackedPath {
     bool tracking_volatility = false;
-    bool erased = false;
     int frame_count = 0;
     SkPath path;
   };
@@ -47,12 +51,7 @@ class VolatilePathTracker {
   // Must be called from the UI task runner.
   //
   // Callers should only insert paths that are currently volatile.
-  void Insert(std::shared_ptr<TrackedPath> path);
-
-  // Removes a path from tracking.
-  //
-  // May be called from any thread.
-  void Erase(std::shared_ptr<TrackedPath> path);
+  void Track(std::shared_ptr<TrackedPath> path);
 
   // Called by the shell at the end of a frame after notifying Dart about idle
   // time.
@@ -67,9 +66,10 @@ class VolatilePathTracker {
 
  private:
   fml::RefPtr<fml::TaskRunner> ui_task_runner_;
-  std::mutex paths_mutex_;
-  std::vector<std::shared_ptr<TrackedPath>> paths_;
+  std::vector<std::weak_ptr<TrackedPath>> paths_;
   bool enabled_ = true;
+
+  friend class testing::ShellTest;
 
   FML_DISALLOW_COPY_AND_ASSIGN(VolatilePathTracker);
 };