@dnfield 2021-12-13
I'm not really sure how we should test this beyond saying it improves an existing benchmark, but I'm open to suggestions. All the changes here are to private implementation details.