@dnfield 2021-12-13
Related: https://github.com/flutter/engine/pull/30293 and https://github.com/flutter/flutter/issues/94585