@dnfield 2021-12-13
Using weak_ptr:

```
2021-12-13T14:28:35-08:00
Running out/host_profile/ui_benchmarks
Run on (16 X 2400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 16384 KiB (x1)
Load Average: 2.55, 2.59, 2.58
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_PlatformMessageResponseDartComplete       1566 us         20.3 us        10000
BM_PathVolatilityTracker                    0.220 ms        0.133 ms         5295
```

Load averages are lower, but even at load averages that are closer it's still faster.