@dnfield 2021-12-14
From CI:

![image](https://user-images.githubusercontent.com/8620741/145915485-b4273205-2544-4aaf-bb19-0916ec25b171.png)

That random up spike was noise